+++ changelog.md
@@ -18,13 +18,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Sorted blog post by descending date
-- Updated home blog section layout and styling
+- Updated home blog section layout, styling, and responsiveness
+- Improved page section padding for better use of whitespace
 
-- Page section height prop
+- Simplified work items styling
+- Simplified page section height prop
+- Simplified page header children prop
+- Simplified about section layout
 
 ### Removed
 
 - Blog section item
+- Team section folder mispelling
+- Loader class
 
 #### 2018-10-22